Move treeMatch into QuoteMatcher
[Cherry-picked 7d3f3e4dced204d6d545d189678c33e02d4df3ba]
